+++ packages/core/src/app/order/OrderConfirmation/OrderConfirmation.tsx
@@ -5,7 +5,7 @@ import {
 import { createRequestSender } from '@bigcommerce/request-sender';
 import React, { type ReactElement, useEffect, useRef, useState } from 'react';
 
-import { useAnalytics, useCheckout } from '@bigcommerce/checkout/contexts';
+import { useAnalytics, useCapabilities, useCheckout } from '@bigcommerce/checkout/contexts';
 import { type ErrorLogger } from '@bigcommerce/checkout/error-handling-utils';
 import { OrderConfirmationPageSkeleton } from '@bigcommerce/checkout/ui';
 
@@ -56,16 +56,17 @@ export const OrderConfirmation = ({
     const embeddedMessengerRef = useRef<EmbeddedCheckoutMessenger | undefined>();
 
     const {
-        checkoutState: {
-            data: { getOrder, getConfig },
-            statuses: { isLoadingOrder },
-        },
+        selectedState: { order, config, isLoadingOrder },
         checkoutService: { loadOrder },
-    } = useCheckout();
+    } = useCheckout(({ data, statuses }) => ({
+        order: data.getOrder(),
+        config: data.getConfig(),
+        isLoadingOrder: statuses.isLoadingOrder(),
+    }));
     const { analyticsTracker } = useAnalytics();
-
-    const config = getConfig();
-    const order = getOrder();
+    const {
+        orderConfirmation: { cannotCreatePersonalAccount },
+    } = useCapabilities();
 
     const handleUnhandledError = (e: Error) => {
         setError(e);
@@ -148,7 +149,7 @@ export const OrderConfirmation = ({
         return <RateLimitedPermalinkView />;
     }
 
-    if (!order || !config || isLoadingOrder()) {
+    if (!order || !config || isLoadingOrder) {
         return <OrderConfirmationPageSkeleton />;
     }
 
@@ -170,6 +171,7 @@ export const OrderConfirmation = ({
 
     return (
         <OrderConfirmationPage
+            cannotCreatePersonalAccount={cannotCreatePersonalAccount}
             currency={currency}
             customerCanBeCreated={customerCanBeCreated}
             error={error}